feat(fdr): added fuel system data to FDR
