+++ fbw-a380x/src/wasm/fbw_a380/src/interface/SimConnectData.h
@@ -2,6 +2,7 @@
 
 #include <MSFS/Legacy/gauges.h>
 #include <SimConnect.h>
+#include "FuelSystemData.h"
 
 struct SimData {
   double nz_g;
@@ -134,309 +135,6 @@ struct SimData {
   double contact_point_compression_2;
   double contact_point_compression_3;
   double contact_point_compression_4;
-};
-
-struct FuelSystemData {
-  double tank_quantity_1_left_outer;
-  double tank_quantity_2_feed1;
-  double tank_quantity_3_left_mid;
-  double tank_quantity_4_left_inner;
-  double tank_quantity_5_feed2;
-  double tank_quantity_6_feed3;
-  double tank_quantity_7_right_inner;
-  double tank_quantity_8_right_mid;
-  double tank_quantity_9_feed4;
-  double tank_quantity_10_right_outer;
-  double tank_quantity_11_trim;
-  double line_1_feed_tank1_to_feed1_tank_pump1;
-  double line_2_feed_tank1_to_feed1_tank_pump2;
-  double line_3_feed_tank2_to_feed2_tank_pump1;
-  double line_4_feed_tank2_to_feed2_tank_pump2;
-  double line_5_feed_tank3_to_feed3_tank_pump1;
-  double line_6_feed_tank3_to_feed3_tank_pump2;
-  double line_7_feed_tank4_to_feed4_tank_pump1;
-  double line_8_feed_tank4_to_feed4_tank_pump2;
-  double line_9_pump1_feed1_to_junc1;
-  double line_10_pump2_feed1_to_junc1;
-  double line_11_pump1_feed2_to_junc2;
-  double line_12_pump2_feed2_to_junc2;
-  double line_13_pump1_feed3_to_junc3;
-  double line_14_pump2_feed3_to_junc3;
-  double line_15_pump1_feed4_to_junc4;
-  double line_16_pump2_feed4_to_junc4;
-  double line_17_junc1_to_eng1lp_valve;
-  double line_18_junc2_to_eng2lp_valve;
-  double line_19_junc3_to_eng3lp_valve;
-  double line_20_junc4_to_eng4lp_valve;
-  double line_21_eng1lp_valve_to_eng1;
-  double line_22_eng2lp_valve_to_eng2;
-  double line_23_eng3lp_valve_to_eng3;
-  double line_24_eng4lp_valve_to_eng4;
-  double line_25_left_outer_tank_to_left_outer_tank_pump;
-  double line_26_left_mid_tank_to_left_mid_tank_pump_fwd;
-  double line_27_left_mid_tank_to_left_mid_tank_pump_aft;
-  double line_28_left_inner_tank_to_left_inner_tank_pump_fwd;
-  double line_29_left_inner_tank_to_left_inner_tank_pump_aft;
-  double line_30_right_outer_tank_to_right_outer_tank_pump;
-  double line_31_right_mid_tank_to_right_mid_tank_pump_fwd;
-  double line_32_right_mid_tank_to_right_mid_tank_pump_aft;
-  double line_33_right_inner_tank_to_right_inner_tank_pump_fwd;
-  double line_34_right_inner_tank_to_right_inner_tank_pump_aft;
-  double line_35_trim_tank_to_trim_tank_pump_left;
-  double line_36_trim_tank_to_trim_tank_pump_right;
-  double line_37_trim_tank_pump_left_to_trim_line_junction1;
-  double line_38_trim_tank_pump_right_to_trim_line_junction1;
-  double line_39_trim_line_junction1_to_trim_line_junction2;
-  double line_40_trim_line_junction1_to_trim_tank_inlet_valve1;
-  double line_41_trim_line_junction2_to_trim_line_iso_valve_fwd;
-  double line_42_trim_line_junction2_to_trim_line_iso_valve_aft_1;
-  double line_43_left_outer_tank_pump_to_fwd_gallery_junction;
-  double line_44_left_mid_tank_pump_fwd_to_fwd_gallery_junction;
-  double line_45_left_mid_tank_pump_aft_to_aft_gallery_junction1;
-  double line_46_left_inner_tank_pump_fwd_to_fwd_gallery_junction;
-  double line_47_left_inner_tank_pump_aft_to_aft_gallery_junction1;
-  double line_48_right_outer_tank_pump_to_fwd_gallery_junction;
-  double line_49_right_mid_tank_pump_fwd_to_fwd_gallery_junction;
-  double line_50_right_mid_tank_pump_aft_to_aft_gallery_junction1;
-  double line_51_right_inner_tank_pump_fwd_to_fwd_gallery_junction;
-  double line_52_right_inner_tank_pump_aft_to_aft_gallery_junction1;
-  double line_53_trim_line_iso_valve_fwd_to_fwd_gallery_junction;
-  double line_54_trim_line_iso_valve_aft_1_to_aft_gallery_junction1;
-  double line_55_fwd_gallery_junction_to_feed_tank1_fwd_xfer_valve1;
-  double line_56_fwd_gallery_junc_to_feed_tanks2_3_junc1;
-  double line_57_feed_tanks2_3_junc1_to_feed_tank2_fwd_xfer_valve1_1;
-  double line_58_feed_tanks2_3_junc1_to_feed_tank3_fwd_xfer_valve1_1;
-  double line_59_feed_tanks2_3_junc1_to_feed_tank2_fwd_xfer_valve1_2;
-  double line_60_feed_tanks2_3_junc1_to_feed_tank3_fwd_xfer_valve1_2;
-  double line_61_fwd_gallery_junction_to_feed_tank4_fwd_xfer_valve1;
-  double line_62_fwd_gallery_junction_to_feed_tank1_fwd_xfer_valve2;
-  double line_63_fwd_gallery_junc_to_feed_tanks2_3_junc2;
-  double line_64_feed_tanks2_3_junc2_to_feed_tank2_fwd_xfer_valve2_1;
-  double line_65_feed_tanks2_3_junc2_to_feed_tank3_fwd_xfer_valve2_1;
-  double line_66_feed_tanks2_3_junc2_to_feed_tank2_fwd_xfer_valve2_2;
-  double line_67_feed_tanks2_3_junc2_to_feed_tank3_fwd_xfer_valve2_2;
-  double line_68_fwd_gallery_junction_to_feed_tank4_fwd_xfer_valve2;
-  double line_69_fwd_gallery_junction_to_left_inner_fwd_xfer_valve;
-  double line_70_fwd_gallery_junction_to_left_mid_fwd_xfer_valve;
-  double line_71_fwd_gallery_junction_to_left_outer_fwd_xfer_valve;
-  double line_72_fwd_gallery_junction_to_right_inner_fwd_xfer_valve;
-  double line_73_fwd_gallery_junction_to_right_mid_fwd_xfer_valve;
-  double line_74_fwd_gallery_junction_to_right_outer_fwd_xfer_valve;
-  double line_75_feed_tank1_fwd_xfer_valve1_to_feed_tank1;
-  double line_76_feed_tank2_fwd_xfer_valve1_1_to_feed_tank2;
-  double line_77_feed_tank3_fwd_xfer_valve1_1_to_feed_tank3;
-  double line_78_feed_tank2_fwd_xfer_valve1_2_to_feed_tank2;
-  double line_79_feed_tank3_fwd_xfer_valve1_2_to_feed_tank3;
-  double line_80_feed_tank4_fwd_xfer_valve1_to_feed_tank4;
-  double line_81_feed_tank1_fwd_xfer_valve2_to_feed_tank1;
-  double line_82_feed_tank2_fwd_xfer_valve2_1_to_feed_tank2;
-  double line_83_feed_tank3_fwd_xfer_valve2_1_to_feed_tank3;
-  double line_84_feed_tank2_fwd_xfer_valve2_2_to_feed_tank2;
-  double line_85_feed_tank3_fwd_xfer_valve2_2_to_feed_tank3;
-  double line_86_feed_tank4_fwd_xfer_valve2_to_feed_tank4;
-  double line_87_left_inner_fwd_xfer_valve_to_left_inner_tank;
-  double line_88_left_mid_fwd_xfer_valve_to_left_mid_tank;
-  double line_89_left_outer_fwd_xfer_valve_to_left_outer_tank;
-  double line_90_right_inner_fwd_xfer_valve_to_right_inner_tank;
-  double line_91_right_mid_fwd_xfer_valve_to_right_mid_tank;
-  double line_92_right_outer_fwd_xfer_valve_to_right_outer_tank;
-  double line_93_aft_gallery_junction2_to_feed_tank1_aft_xfer_valve1;
-  double line_94_aft_gallery_junction2_to_feed_tank2_aft_xfer_valve1;
-  double line_95_aft_gallery_junction2_to_feed_tank3_aft_xfer_valve1;
-  double line_96_aft_gallery_junction2_to_feed_tank4_aft_xfer_valve1;
-  double line_97_feed_tank1_aft_xfer_valve1_to_feed_tank1_aft_xfer_valve2;
-  double line_98_feed_tank2_aft_xfer_valve1_to_feed_tank2_aft_xfer_valve2;
-  double line_99_feed_tank3_aft_xfer_valve1_to_feed_tank3_aft_xfer_valve2;
-  double line_100_feed_tank4_aft_xfer_valve1_to_feed_tank4_aft_xfer_valve2;
-  double line_101_aft_gallery_junction1_to_left_inner_aft_xfer_valve1;
-  double line_102_aft_gallery_junction1_to_left_mid_aft_xfer_valve1;
-  double line_103_aft_gallery_junction1_to_left_outer_aft_xfer_valve1;
-  double line_104_aft_gallery_junction1_to_right_inner_aft_xfer_valve1;
-  double line_105_aft_gallery_junction1_to_right_mid_aft_xfer_valve1;
-  double line_106_aft_gallery_junction1_to_right_outer_aft_xfer_valve1;
-  double line_107_left_inner_aft_transfer_valve1_to_left_inner_aft_xfer_valve2;
-  double line_108_left_mid_aft_transfer_valve1_to_left_mid_aft_xfer_valve2;
-  double line_109_left_outer_aft_transfer_valve1_to_left_outer_aft_xfer_valve2;
-  double line_110_right_inner_aft_transfer_valve1_to_right_inner_aft_xfer_valve2;
-  double line_111_right_mid_aft_transfer_valve1_to_right_mid_aft_xfer_valve2;
-  double line_112_right_outer_aft_transfer_valve1_to_right_outer_aft_xfer_valve2;
-  double line_113_feed_tank1_aft_xfer_valve2_to_feed_tank1;
-  double line_114_feed_tank2_aft_xfer_valve2_to_feed_tank2;
-  double line_115_feed_tank3_aft_xfer_valve2_to_feed_tank3;
-  double line_116_feed_tank4_aft_xfer_valve2_to_feed_tank4;
-  double line_117_left_inner_aft_xfer_valve2_to_left_inner_tank;
-  double line_118_left_mid_aft_xfer_valve2_to_left_mid_tank;
-  double line_119_left_outer_aft_xfer_valve2_to_left_outer_tank;
-  double line_120_right_inner_aft_xfer_valve2_to_right_inner_tank;
-  double line_121_right_mid_aft_xfer_valve2_to_right_mid_tank;
-  double line_122_right_outer_aft_xfer_valve2_to_right_outer_tank;
-  double line_123_fwd_gallery_junc_to_gallery_aux_revalve_left;
-  double line_124_fwd_gallery_junc_to_gallery_aux_revalve_right;
-  double line_125_gallery_aux_revalve_left_to_a_ft_gallery_junc1;
-  double line_126_gallery_aux_revalve_right_to_a_ft_gallery_junc1;
-  double line_127_fwd_gallery_junc_to_transfer_devalve;
-  double line_128_junc1_to_cross_feed_valve1;
-  double line_129_junc2_to_cross_feed_valve2;
-  double line_130_junc3_to_cross_feed_valve3;
-  double line_131_junc4_to_cross_feed_valve4;
-  double line_132_cross_feed_valve1_to_cross_feed_junc1;
-  double line_133_cross_feed_junc1_to_cross_feed_valve2;
-  double line_134_cross_feed_junc1_to_cross_feed_junc2;
-  double line_135_cross_feed_junc2_to_cross_feed_valve3;
-  double line_136_cross_feed_junc2_to_cross_feed_valve4;
-  double line_137_transfer_devalve_to_cross_feed_junc1;
-  double line_138_junc4_to_apu_feed_pump;
-  double line_139_apu_feed_pump_to_apu_iso_valve;
-  double line_140_apu_iso_valve_to_apulp_valve;
-  double line_141_apulp_valve_to_apu;
-  double line_142_left_outer_tank_to_left_outer_emer_xfer_valve;
-  double line_143_right_outer_tank_to_right_outer_emer_xfer_valve;
-  double line_144_left_outer_emer_xfer_valve_to_feed_tank1;
-  double line_145_right_outer_emer_xfer_valve_to_feed_tank4;
-  double line_146_aft_gallery_junction1_to_jetisson_nozzle_valve_left;
-  double line_147_aft_gallery_junction1_to_jetisson_nozzle_valve_right;
-  double line_148_trim_tank_inlet_valve1_to_trim_tank;
-  double line_149_trim_tank_inlet_valve2_to_trim_tank;
-  double line_150_trim_line_junction2_to_trim_line_iso_valve_aft_2;
-  double line_151_trim_line_junction1_to_trim_tank_inlet_valve2;
-  double line_152_trim_line_iso_valve_aft_2_to_aft_gallery_junction1;
-  double line_153_aft_gallery_junction1_to_aft_gallery_junction2;
-  double junction_1_junction1;
-  double junction_2_junction2;
-  double junction_3_junction3;
-  double junction_4_junction4;
-  double junction_5_trim_line_junction1;
-  double junction_6_trim_line_junction2;
-  double junction_7_fwd_gallery_junction;
-  double junction_8_feed_tanks2_3_junction1;
-  double junction_9_feed_tanks2_3_junction2;
-  double junction_10_aft_gallery_junction1;
-  double junction_11_aft_gallery_junction2;
-  double junction_12_cross_feed_junc;
-  double junction_13_cross_feed_junc;
-  double valve_1_engine1lp_valve;
-  double valve_2_engine2lp_valve;
-  double valve_3_engine3lp_valve;
-  double valve_4_engine4lp_valve;
-  double valve_5_feed_tank1_fwd_transfer_valve1;
-  double valve_6_feed_tank2_fwd_transfer_valve1_1;
-  double valve_7_feed_tank3_fwd_transfer_valve1_1;
-  double valve_8_feed_tank2_fwd_transfer_valve1_2;
-  double valve_9_feed_tank3_fwd_transfer_valve1_2;
-  double valve_10_feed_tank4_fwd_transfer_valve1;
-  double valve_11_feed_tank1_fwd_transfer_valve2;
-  double valve_12_feed_tank2_fwd_transfer_valve2_1;
-  double valve_13_feed_tank3_fwd_transfer_valve2_1;
-  double valve_14_feed_tank2_fwd_transfer_valve2_2;
-  double valve_15_feed_tank3_fwd_transfer_valve2_2;
-  double valve_16_feed_tank4_fwd_transfer_valve2;
-  double valve_17_left_inner_fwd_transfer_valve;
-  double valve_18_left_mid_fwd_transfer_valve;
-  double valve_19_left_outer_fwd_transfer_valve;
-  double valve_20_right_inner_fwd_transfer_valve;
-  double valve_21_right_mid_fwd_transfer_valve;
-  double valve_22_right_outer_fwd_transfer_valve;
-  double valve_23_feed_tank1_aft_transfer_valve1;
-  double valve_24_feed_tank2_aft_transfer_valve1;
-  double valve_25_feed_tank3_aft_transfer_valve1;
-  double valve_26_feed_tank4_aft_transfer_valve1;
-  double valve_27_feed_tank1_aft_transfer_valve2;
-  double valve_28_feed_tank2_aft_transfer_valve2;
-  double valve_29_feed_tank3_aft_transfer_valve2;
-  double valve_30_feed_tank4_aft_transfer_valve2;
-  double valve_31_left_inner_aft_transfer_valve1;
-  double valve_32_left_mid_aft_transfer_valve1;
-  double valve_33_left_outer_aft_transfer_valve1;
-  double valve_34_right_inner_aft_transfer_valve1;
-  double valve_35_right_mid_aft_transfer_valve1;
-  double valve_36_right_outer_aft_transfer_valve1;
-  double valve_37_left_inner_aft_transfer_valve2;
-  double valve_38_left_mid_aft_transfer_valve2;
-  double valve_39_left_outer_aft_transfer_valve2;
-  double valve_40_right_inner_aft_transfer_valve2;
-  double valve_41_right_mid_aft_transfer_valve2;
-  double valve_42_right_outer_aft_transfer_valve2;
-  double valve_43_trim_tank_inlet_valve1;
-  double valve_44_trim_line_isolation_valve_fwd;
-  double valve_45_trim_line_isolation_valve_aft_1;
-  double valve_46_cross_feed_valve1;
-  double valve_47_cross_feed_valve2;
-  double valve_48_cross_feed_valve3;
-  double valve_49_cross_feed_valve4;
-  double valve_50_apu_iso_valve;
-  double valve_51_apulp_valve;
-  double valve_52_left_outer_emer_transfer_valve;
-  double valve_53_right_outer_emer_transfer_valve;
-  double valve_54_gallery_aux_refuel_valve_left;
-  double valve_55_gallery_aux_refuel_valve_right;
-  double valve_56_transfer_defuel_valve;
-  double valve_57_jettison_nozzle_valve_left;
-  double valve_58_jettison_nozzle_valve_right;
-  double valve_59_trim_line_isolation_valve_aft_2;
-  double valve_60_trim_tank_inlet_valve2;
-  double pump_1_feed1_tank_pump1;
-  double pump_2_feed1_tank_pump2;
-  double pump_3_feed2_tank_pump1;
-  double pump_4_feed2_tank_pump2;
-  double pump_5_feed3_tank_pump1;
-  double pump_6_feed3_tank_pump2;
-  double pump_7_feed4_tank_pump1;
-  double pump_8_feed4_tank_pump2;
-  double pump_9_left_outer_tank_pump;
-  double pump_10_left_mid_tank_pump_fwd;
-  double pump_11_left_mid_tank_pump_aft;
-  double pump_12_left_inner_tank_pump_fwd;
-  double pump_13_right_inner_tank_pump_fwd;
-  double pump_14_right_outer_tank_pump;
-  double pump_15_right_mid_tank_pump_fwd;
-  double pump_16_right_mid_tank_pump_aft;
-  double pump_17_left_inner_tank_pump_aft;
-  double pump_18_right_inner_tank_pump_aft;
-  double pump_19_trim_tank_pump_left;
-  double pump_20_trim_tank_pump_right;
-  double pump_21_apu_feed_pump;
-  double trigger_1_innerand_mid_tanks_xfer_feed1_start;
-  double trigger_2_innerand_mid_tanks_xfer_feed2_start;
-  double trigger_3_innerand_mid_tanks_xfer_feed3_start;
-  double trigger_4_innerand_mid_tanks_xfer_feed4_start;
-  double trigger_5_equalize1and4;
-  double trigger_6_equalize2and3;
-  double trigger_7_innerand_mid_tanks_xfer_feed1and4_end;
-  double trigger_8_innerand_mid_tanks_xfer_feed2and3_end;
-  double trigger_9_innerand_mid_tanks_xfer_feed2and3_end_redundant;
-  double trigger_10_innerand_mid_tanks_xfer_feed1and4_end_redundant;
-  double trigger_11_inner_tank_left_empty;
-  double trigger_12_inner_tank_right_empty;
-  double trigger_13_mid_tanks_below8000;
-  double trigger_14_mid_below8000_threshold_feed2_start;
-  double trigger_15_mid_below8000_threshold_feed3_start;
-  double trigger_16_mid_below8000_threshold_feed2_end;
-  double trigger_17_mid_below8000_threshold_feed3_end;
-  double trigger_18_equalize1and3for_mid_below8000;
-  double trigger_19_equalize1and2for_mid_below8000;
-  double trigger_20_equalize2and4for_mid_below8000;
-  double trigger_21_equalize3and4for_mid_below8000;
-  double trigger_22_mid_tank_left_empty;
-  double trigger_23_mid_tank_right_empty;
-  double trigger_24_trim_tank_transfer_to_feed_tank1;
-  double trigger_25_trim_tank_transfer_to_feed_tank2;
-  double trigger_26_trim_tank_transfer_to_feed_tank3;
-  double trigger_27_trim_tank_transfer_to_feed_tank4;
-  double trigger_28_equalize1and3;
-  double trigger_29_equalize1and2;
-  double trigger_30_equalize2and4;
-  double trigger_31_equalize3and4;
-  double trigger_32_trim_tank_empty;
-  double trigger_33_outer_tanks_transfer_to_feed_tank1;
-  double trigger_34_outer_tanks_transfer_to_feed_tank2;
-  double trigger_35_outer_tanks_transfer_to_feed_tank1;
-  double trigger_36_outer_tanks_transfer_to_feed_tank2;
-  double trigger_37_outer_tanks_transfer_to_feed_tank1_end;
-  double trigger_38_outer_tanks_transfer_to_feed_tank2_end;
-  double trigger_39_outer_tanks_transfer_to_feed_tank3_end;
-  double trigger_40_outer_tanks_transfer_to_feed_tank4_end;
-  double trigger_41_cg_control_transfer_start;
-  double trigger_42_cg_control_transfer_end;
 };
 
 struct SimInput {
