+++ tools/fdr2csv/a380x/wrapper.hpp
@@ -1,3 +1,4 @@
+#include "../../../fbw-a380x/src/wasm/fbw_a380/src/interface/FuelSystemData.h"
 #include "../../../fbw-a380x/src/wasm/fbw_a380/src/model/A380PrimComputer_types.h"
 #include "../../../fbw-a380x/src/wasm/fbw_a380/src/model/A380SecComputer_types.h"
 #include "../../../fbw-a380x/src/wasm/fbw_a380/src/model/AutopilotLaws_types.h"
